Allow dynamic switching between un/managed blocks
Boehm supports both collectable and uncollectable allocations. The
latter being blocks which are still allocated by the GC and scanned for
pointers, but are only deallocated via explicit calls to `GC_free`. Once
an object is allocated, its allocation kind is fixed. In other words,
there is no mechanism to dynamically convert an object kind from
collectable to uncollectable or vice versa.

Allowing this kind of dynamic switching is not really feasible because
of the way the allocator is designed. Objects are allocated in distinct
pools based on their size-class and kind. These pools keep objects of
the same kind in the same page and share page-level metadata needed for
a GC. For the collector to behave correctly, changing an object's kind
would require memmoving it to a different page which is too expensive.

This commit addresses this by splitting the `collectable` allocation
kind into two sub-kinds: `managed`, and `unmanaged`. Objects allocated
with `GC_malloc` are pinned in memory until they are converted to
`managed` using the `GC_set_managed` API.

This change is only possible with the `USE_MARK_BYTES` flag enabled,
which means that each object uses 1 machine word for its mark state,
instead of a page-level marking bitmap. Fortunately this is already
enabled by default, since its main purpose is to reduce atomic
load/store pressure on the mark state when using parallel marker
threads.

We use the 2nd bit in the mark word as an is_managed flag. This means
that in most cases we don't actually introduce any additional branches
in fast paths of the collector because it already requires bitmasking to
do a mark check. However, I'm not yet sure of the broader implications
this has on performance because it will lead to greater freelist
fragmentation in the collectable pool, and could pessimise generational
GC due to fewer objects dying young.