Feature gate allocation switching
This can be enabled with the configure flag --enable-alloc-switching or
the CMake arg -Denable-allocation-switching=ON.

Feature gating this means we can disable the uncollectable pool entirely
in Boehm. This pool relied on lazily clearing the mark bit, which is no
longer the case the introducion of an additional managed bit in 0f0cced.
This caused a huge slowdown when allocating lots of uncollectable
objects because they were erroneously enqueued for finalization.

No longer using the uncollectable pool also means several branches in
allocation and marking fast paths can be removed.