Add a null check before pushing TLS roots on scheduling thread
Occasionally a GC can be scheduled during the construction of a thread.
This is fine, because Boehm handles this by simply not pushing that
thread's stack to the marking worklist as it won't yet contain any
roots.

However, this does mean that the thread's context may not yet have been
initialised. Therefore, before we push the TLS roots to its context we
must first perform a null check.