Flesh out API for malloc with buffered finalization
This introduces a new malloc function `GC_buffered_finalize_malloc` which
ensures that unreachable objects are added to a 4KiB thread-local buffer to
be later finalized. Once the buffer is full, it's passed to a user supplied
closure.

The disclaim API in Boehm has no mechanism for detecting which
allocations in a free-list are simply empty and which require finalizing.
So in order to get around this we change mark procedure, using an extra bit
in the object header to determine whether or not an object has a
finalization closure.