Use separate thread for finalisation
This commit creates a finalisation thread in the GC_INIT() routine. The
thread spins indefinitely waiting to process objects in the finaliser
buffers.

This commit also makes sure that once an object has been finalised, its
finalised bit is set so that it is not re-added to the buffer the next
time it becomes unreachable.