Mark TLS roots of scheduling GC thread
When a mutator thread schedules a GC, it suspends all other mutator
threads by sending SIGPWR signals. Before those other threads are
suspended, they push a pointer to their TLS roots to the collector so
their TLS roots can be marked.

Since this happened in the suspend signal handler, the TLS roots of the
thread that invoked GC are never marked (because the scheduling thread
is never suspended!) This fixes that by ensuring the TLS roots of the
scheduling thread are pushed when the GC enters the stop the world
phase.