Revert to using old GC_finalize mechanism
Using an extension of the disclaim API proved difficult because of the
need to re-mark objects pointed to by finalisable objects before a
finaliser is run. This would have led to complex changes to the
mark-phase for no real benefit.

Instead, by switching back to the GC_finalize mechanism but ensuring
that finalisers are run off of mutator threads, we get this desired
behaviour 'for free'.

This approach uses a simple condition variable to sleep the finaliser thread
when there is no finalisation work to do. This is then potentially
resumed at the end of a GC cycle if that cycle discovered finalisable
objects.