Add tag for tracking if object is RAII'd or GC'd
To support finalizer elison in Alloy, we need a new way to differentiate
between RAII (non-GC'd) and GC'd objects such that this status can be
changed dynamically. Currently in BDWGC, if we use GC_malloc and
GC_malloc_uncollectable for this distinction, there is no way to change
this after an object has been allocated. Using a spare bit in the mark
header is a convenient way for us to tell the sweep phase whether or not
an object should be considered garbage or whether it is uncollectable
while allowing it to be changed in the future.