Add API for tracking reclaimed objects
This is needed in order to track how many things in a program are
actually 'managed' by the collector -- i.e. `Gc<T>s` and any heap
allocated components of `T` which are transitively owned by the `Gc<T>`
but not directly allocated by it.

Without this API, we used a hodgepodge of recorded leaks, elided
finalizers, ratios to previously converted programs, etc to gain a rough
estimate for this figure. This turned out to be pretty inaccurate and I
didn't trust the numbers.